+++ src/payload.config.ts
@@ -35,10 +35,10 @@ import { Page, Post } from 'src/payload-types'
 
 import { searchFields } from '@/search/fieldOverrides'
 import { beforeSyncWithSearch } from '@/search/beforeSync'
-import { Blogposts } from "@/collections/Blogposts/Blogposts";
+import { Blogposts } from "@/collections/Blogposts";
 import { Authors } from "@/collections/Authors";
 import { CaseStudies } from "@/collections/CaseStudies";
-import { Podcasts } from "@/collections/Podcasts/Podcasts";
+import { Podcasts } from "@/collections/Podcasts";
 import { TalksAndRoundtables } from "@/collections/TalksAndRoundtables";
 import { HomePageSettings } from "@/Globals/HubHighlights/config";
 import { cloudStoragePlugin } from "@payloadcms/plugin-cloud-storage";
